+++ src/routes/userRoutes.js
@@ -1,0 +1,9 @@
+//* Import express and initialize the routers
+import express from "express";
+const router = express.Router();
+
+import { getUser } from "../controllers/userController";
+
+router.get("/", getUser);
+
+export default router;
